+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelDataSource.swift
@@ -38,6 +38,9 @@ protocol ChannelDataSource: AnyObject {
 
     /// List of the messages.
     var messages: LazyCachedMapCollection<ChatMessage> { get }
+    
+    /// Determines whether all new messages have been fetched.
+    var hasLoadedAllNextMessages: Bool { get }
 
     /// Loads the previous messages.
     /// - Parameters:
@@ -49,6 +52,28 @@ protocol ChannelDataSource: AnyObject {
         limit: Int,
         completion: ((Error?) -> Void)?
     )
+    
+    /// Loads newer messages.
+    /// - Parameters:
+    ///  - limit: the max number of messages to be retrieved.
+    ///  - completion: called when the messages are loaded.
+    func loadNextMessages(
+        limit: Int,
+        completion: ((Error?) -> Void)?
+    )
+    
+    /// Loads a page around the provided message id.
+    /// - Parameters:
+    ///  - messageId: the id of the message.
+    ///  - completion: called when the messages are loaded.
+    func loadPageAroundMessageId(
+        _ messageId: MessageId,
+        completion: ((Error?) -> Void)?
+    )
+    
+    /// Loads the first page of the channel.
+    ///  - Parameter completion: called when the initial page is loaded.
+    func loadFirstPage(_ completion: ((_ error: Error?) -> Void)?)
 }
 
 /// Implementation of `ChannelDataSource`. Loads the messages of the channel.
@@ -56,8 +81,13 @@ class ChatChannelDataSource: ChannelDataSource, ChatChannelControllerDelegate {
 
     let controller: ChatChannelController
     weak var delegate: MessagesDataSource?
+    
     var messages: LazyCachedMapCollection<ChatMessage> {
         controller.messages
+    }
+    
+    var hasLoadedAllNextMessages: Bool {
+        controller.hasLoadedAllNextMessages
     }
 
     init(controller: ChatChannelController) {
@@ -98,6 +128,21 @@ class ChatChannelDataSource: ChannelDataSource, ChatChannelControllerDelegate {
             completion: completion
         )
     }
+    
+    func loadNextMessages(limit: Int, completion: ((Error?) -> Void)?) {
+        controller.loadNextMessages(limit: limit, completion: completion)
+    }
+    
+    func loadPageAroundMessageId(
+        _ messageId: MessageId,
+        completion: ((Error?) -> Void)?
+    ) {
+        controller.loadPageAroundMessageId(messageId, completion: completion)
+    }
+    
+    func loadFirstPage(_ completion: ((_ error: Error?) -> Void)?) {
+        controller.loadFirstPage(completion)
+    }
 }
 
 /// Implementation of the `ChannelDataSource`. Loads the messages in a reply thread.
@@ -105,9 +150,15 @@ class MessageThreadDataSource: ChannelDataSource, ChatMessageControllerDelegate 
 
     let channelController: ChatChannelController
     let messageController: ChatMessageController
+    
     weak var delegate: MessagesDataSource?
+    
     var messages: LazyCachedMapCollection<ChatMessage> {
         messageController.replies
+    }
+    
+    var hasLoadedAllNextMessages: Bool {
+        messageController.hasLoadedAllNextReplies
     }
 
     init(
@@ -159,5 +210,20 @@ class MessageThreadDataSource: ChannelDataSource, ChatMessageControllerDelegate 
             limit: limit,
             completion: completion
         )
+    }
+    
+    func loadNextMessages(limit: Int, completion: ((Error?) -> Void)?) {
+        messageController.loadNextReplies(limit: limit, completion: completion)
+    }
+    
+    func loadPageAroundMessageId(
+        _ messageId: MessageId,
+        completion: ((Error?) -> Void)?
+    ) {
+        messageController.loadPageAroundReplyId(messageId, completion: completion)
+    }
+    
+    func loadFirstPage(_ completion: ((_ error: Error?) -> Void)?) {
+        messageController.loadFirstPage(completion)
     }
 }