+++ Sources/StreamChatSwiftUI/ChatChannel/ChatChannelView.swift
@@ -55,14 +55,17 @@ public struct ChatChannelView<Factory: ViewFactory>: View, KeyboardReadable {
                             listId: viewModel.listId,
                             isMessageThread: viewModel.isMessageThread,
                             shouldShowTypingIndicator: viewModel.shouldShowTypingIndicator,
-                            onMessageAppear: viewModel.handleMessageAppear(index:),
+                            scrollPosition: $viewModel.scrollPosition,
+                            loadingNextMessages: viewModel.loadingNextMessages,
+                            onMessageAppear: viewModel.handleMessageAppear(index:scrollDirection:),
                             onScrollToBottom: viewModel.scrollToLastMessage,
                             onLongPress: { displayInfo in
                                 messageDisplayInfo = displayInfo
                                 withAnimation {
                                     viewModel.showReactionOverlay(for: AnyView(self))
                                 }
-                            }
+                            },
+                            onJumpToMessage: viewModel.jumpToMessage(messageId:)
                         )
                         .overlay(
                             viewModel.currentDateString != nil ?