+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListView.swift
@@ -18,18 +18,22 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
     @Binding var scrolledId: String?
     @Binding var showScrollToLatestButton: Bool
     @Binding var quotedMessage: ChatMessage?
+    @Binding var scrollPosition: String?
+    var loadingNextMessages: Bool
     var currentDateString: String?
     var listId: String
     var isMessageThread: Bool
     var shouldShowTypingIndicator: Bool
-
-    var onMessageAppear: (Int) -> Void
+    
+    var onMessageAppear: (Int, ScrollDirection) -> Void
     var onScrollToBottom: () -> Void
     var onLongPress: (MessageDisplayInfo) -> Void
-
+    var onJumpToMessage: ((String) -> Bool)?
+    
     @State private var width: CGFloat?
     @State private var keyboardShown = false
     @State private var pendingKeyboardUpdate: Bool?
+    @State private var scrollDirection = ScrollDirection.up
     @State private var newMessagesStartId: String?
 
     private var messageRenderingUtil = MessageRenderingUtil.shared
@@ -69,9 +73,12 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
         listId: String,
         isMessageThread: Bool = false,
         shouldShowTypingIndicator: Bool = false,
-        onMessageAppear: @escaping (Int) -> Void,
+        scrollPosition: Binding<String?> = .constant(nil),
+        loadingNextMessages: Bool = false,
+        onMessageAppear: @escaping (Int, ScrollDirection) -> Void,
         onScrollToBottom: @escaping () -> Void,
-        onLongPress: @escaping (MessageDisplayInfo) -> Void
+        onLongPress: @escaping (MessageDisplayInfo) -> Void,
+        onJumpToMessage: ((String) -> Bool)? = nil
     ) {
         self.factory = factory
         self.channel = channel
@@ -83,10 +90,13 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
         self.onMessageAppear = onMessageAppear
         self.onScrollToBottom = onScrollToBottom
         self.onLongPress = onLongPress
+        self.onJumpToMessage = onJumpToMessage
         self.shouldShowTypingIndicator = shouldShowTypingIndicator
+        self.loadingNextMessages = loadingNextMessages
         _scrolledId = scrolledId
         _showScrollToLatestButton = showScrollToLatestButton
         _quotedMessage = quotedMessage
+        _scrollPosition = scrollPosition
         if !messageRenderingUtil.hasPreviousMessageSet
             || self.showScrollToLatestButton == false
             || self.scrolledId != nil
@@ -144,7 +154,7 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                                     index = messageListDateUtils.index(for: message, in: messages)
                                 }
                                 if let index = index {
-                                    onMessageAppear(index)
+                                    onMessageAppear(index, scrollDirection)
                                 }
                             }
                             .padding(
@@ -189,7 +199,9 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                         .id(listId)
                     }
                     .modifier(factory.makeMessageListModifier())
+                    .modifier(ScrollTargetLayoutModifier(enabled: loadingNextMessages))
                 }
+                .modifier(ScrollPositionModifier(scrollPosition: loadingNextMessages ? $scrollPosition : .constant(nil)))
                 .background(
                     factory.makeMessageListBackground(
                         colors: colors,
@@ -206,6 +218,15 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                     DispatchQueue.main.async {
                         let offsetValue = value ?? 0
                         let diff = offsetValue - utils.messageCachingUtils.scrollOffset
+                        if abs(diff) > 15 {
+                            if diff > 0 {
+                                if scrollDirection == .up {
+                                    scrollDirection = .down
+                                }
+                            } else if diff < 0 && scrollDirection == .down {
+                                scrollDirection = .up
+                            }
+                        }
                         utils.messageCachingUtils.scrollOffset = offsetValue
                         let scrollButtonShown = offsetValue < -20
                         if scrollButtonShown != showScrollToLatestButton {
@@ -215,6 +236,9 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                             keyboardShown = false
                             resignFirstResponder()
                         }
+                        if offsetValue > 5 {
+                            onMessageAppear(0, .down)
+                        }
                     }
                 }
                 .flippedUpsideDown()
@@ -222,12 +246,9 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
                 .clipped()
                 .onChange(of: scrolledId) { scrolledId in
                     if let scrolledId = scrolledId {
-                        if scrolledId == messages.first?.messageId {
-                            self.scrolledId = nil
-                        } else {
-                            DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
-                                self.scrolledId = nil
-                            }
+                        let shouldJump = onJumpToMessage?(scrolledId) ?? false
+                        if !shouldJump {
+                            return
                         }
                         withAnimation {
                             scrollView.scrollTo(scrolledId, anchor: messageListConfig.scrollingAnchor)
@@ -343,6 +364,42 @@ public struct MessageListView<Factory: ViewFactory>: View, KeyboardReadable {
             onLongPress(messageDisplayInfo)
         }
     }
+}
+
+struct ScrollPositionModifier: ViewModifier {
+    @Binding var scrollPosition: String?
+    
+    func body(content: Content) -> some View {
+        if #available(iOS 17, *) {
+            content
+                .scrollPosition(id: $scrollPosition, anchor: .top)
+        } else {
+            content
+        }
+    }
+}
+
+struct ScrollTargetLayoutModifier: ViewModifier {
+    
+    var enabled: Bool
+    
+    func body(content: Content) -> some View {
+        if !enabled {
+            return content
+        }
+        if #available(iOS 17, *) {
+            return content
+                .scrollTargetLayout(isEnabled: enabled)
+                .scrollTargetBehavior(.paging)
+        } else {
+            return content
+        }
+    }
+}
+
+public enum ScrollDirection {
+    case up
+    case down
 }
 
 public struct NewMessagesIndicator: View {